requirements.txt
pandas
openpyxl
pypdf
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
streamlit
plotly
google-generativeai
python-dotenv
langchain-huggingface
langchain-chroma
langchain-ollama
langchain-xai
langchain-groq
torch
huggingface-hub
